semi-working prevalence plot
